+++ models/StorageFile.php
@@ -25,8 +25,6 @@ use yii\base\Event;
  * This is the model class for table "{{%cms_storage_file}}".
  *
  * @property integer $id
- * @property string $src
- * @property string $absoluteSrc
  * @property string $cluster_id
  * @property string $cluster_file
  * @property integer $created_by
@@ -34,7 +32,6 @@ use yii\base\Event;
  * @property integer $created_at
  * @property integer $updated_at
  * @property string $size
- * @property string $type
  * @property string $mime_type
  * @property string $extension
  * @property string $original_name
@@ -44,12 +41,11 @@ use yii\base\Event;
  * @property string $description_full
  * @property integer $image_height
  * @property integer $image_width
- * @property string $linked_to_model @version > 2.4.*  is depricated
- * @property string $linked_to_value @version > 2.4.*  is depricated
- * @property string $rootSrc
  *
- * @property User $updatedBy
- * @property User $createdBy
+ * @property string $src
+ * @property string $absoluteSrc
+ *
+ * @property \skeeks\cms\components\storage\Cluster $cluster
  */
 class StorageFile extends Core
 {
@@ -67,13 +63,11 @@ class StorageFile extends Core
     public function rules()
     {
         return array_merge(parent::rules(), [
-            [['src'], 'required'],
-            [['created_by', 'updated_by', 'created_at', 'updated_at', 'size', 'image_height', 'image_width', 'published_at'], 'integer'],
+            [['created_by', 'updated_by', 'created_at', 'updated_at', 'size', 'image_height', 'image_width'], 'integer'],
             [['description_short', 'description_full'], 'string'],
-            [['src', 'cluster_file', 'original_name', 'name', 'linked_to_model', 'linked_to_value'], 'string', 'max' => 255],
-            [['cluster_id', 'type', 'mime_type', 'extension'], 'string', 'max' => 16],
+            [['cluster_file', 'original_name', 'name'], 'string', 'max' => 255],
+            [['cluster_id', 'mime_type', 'extension'], 'string', 'max' => 16],
             [['name_to_save'], 'string', 'max' => 32],
-            [['src'], 'unique'],
             [['cluster_id', 'cluster_file'], 'unique', 'targetAttribute' => ['cluster_id', 'cluster_file'], 'message' => Yii::t('app','The combination of Cluster ID and Cluster Src has already been taken.')],
         ]);
     }
@@ -85,7 +79,6 @@ class StorageFile extends Core
     {
         return array_merge(parent::attributeLabels(), [
             'id' => Yii::t('app', 'ID'),
-            'src' => Yii::t('app', 'Src'),
             'cluster_id' => Yii::t('app', 'Storage'),
             'cluster_file' => Yii::t('app', 'Cluster File'),
             'created_by' => Yii::t('app', 'Created By'),
@@ -93,7 +86,6 @@ class StorageFile extends Core
             'created_at' => Yii::t('app', 'Created At'),
             'updated_at' => Yii::t('app', 'Updated At'),
             'size' => Yii::t('app', 'File Size'),
-            'type' => Yii::t('app', 'Type'),
             'mime_type' => Yii::t('app', 'File Type'),
             'extension' => Yii::t('app', 'Extension'),
             'original_name' => Yii::t('app', 'Original FileName'),
@@ -103,9 +95,6 @@ class StorageFile extends Core
             'description_full' => Yii::t('app', 'Description Full'),
             'image_height' => Yii::t('app', 'Image Height'),
             'image_width' => Yii::t('app', 'Image Width'),
-            'linked_to_model' => Yii::t('app', 'Linked To Model'),
-            'linked_to_value' => Yii::t('app', 'Linked To Value'),
-            'published_at'  => Yii::t('app', 'Published At'),
         ]);
     }
 
@@ -113,16 +102,6 @@ class StorageFile extends Core
     const TYPE_FILE     = "file";
     const TYPE_IMAGE    = "image";
 
-
-    /**
-     * @inheritdoc
-     */
-    public function behaviors()
-    {
-        return array_merge(parent::behaviors(), [
-            TimestampPublishedBehavior::className() => TimestampPublishedBehavior::className()
-        ]);
-    }
 
     /**
      * @return bool|int
@@ -133,7 +112,7 @@ class StorageFile extends Core
         //Сначала удалить файл
         try
         {
-            $cluster = $this->cluster();
+            $cluster = $this->cluster;
 
             $cluster->deleteTmpDir($this->cluster_file);
             $cluster->delete($this->cluster_file);
@@ -151,8 +130,7 @@ class StorageFile extends Core
      */
     public function deleteTmpDir()
     {
-        $cluster = $this->cluster();
-        $cluster->deleteTmpDir($this->cluster_file);
+        $this->cluster->deleteTmpDir($this->cluster_file);
 
         return $this;
     }
@@ -182,14 +160,6 @@ class StorageFile extends Core
     }
 
     /**
-     * @return \skeeks\cms\components\storage\Cluster
-     */
-    public function cluster()
-    {
-        return \Yii::$app->storage->getCluster($this->cluster_id);
-    }
-
-    /**
      * TODO: Переписать нормально
      * Обновление информации о файле
      *
@@ -199,7 +169,7 @@ class StorageFile extends Core
     {
         $src = $this->src;
 
-        if ($this->cluster() instanceof ClusterLocal)
+        if ($this->cluster instanceof ClusterLocal)
         {
             if (!\Yii::$app->request->hostInfo)
             {
@@ -229,30 +199,45 @@ class StorageFile extends Core
      */
     public function getAbsoluteSrc()
     {
-        return $this->cluster()->getAbsoluteUrl($this->cluster_file);
+        return $this->cluster->getAbsoluteUrl($this->cluster_file);
     }
 
     /**
+     * Путь к файлу
+     * @return string
+     */
+    public function getSrc()
+    {
+        return $this->cluster->getPublicSrc($this->cluster_file);
+    }
+
+
+    /**
+     * @return \skeeks\cms\components\storage\Cluster
+     */
+    public function getCluster()
+    {
+        return \Yii::$app->storage->getCluster($this->cluster_id);
+    }
+
+
+    /**
+     * TODO::is depricated version > 2.6.0
+     * @return \skeeks\cms\components\storage\Cluster
+     */
+    public function cluster()
+    {
+        return $this->cluster;
+    }
+
+    /**
+     * TODO::is depricated version > 2.6.0
      * @return string
      */
     public function getRootSrc()
     {
-        return $this->cluster()->getRootSrc($this->cluster_file);
+        return $this->cluster->getRootSrc($this->cluster_file);
     }
-
-    /*public function getSrc()
-    {
-        //echo $this->cluster_file;die;
-        return $this->cluster()->getPublicSrc($this->cluster_file);
-    }*/
-
-    /**
-     * @inheritdoc
-     */
-    /*public static function findIdentity($src)
-    {
-        return static::findOne(['src' => $src]);
-    }*/
 }