+++ widgets/StorageFileManager.php
@@ -122,13 +122,6 @@ HTML;
 
         ];
 
-        if ($this->model)
-        {
-            /*$urlData = ArrayHelper::merge($urlData, [
-                "linked_to_model"   => $this->model->getRef()->getCode(),
-                "linked_to_value"   => $this->model->getRef()->getValue(),
-            ]);*/
-        }
 
         if ($this->fileGroup)
         {
@@ -162,15 +155,6 @@ HTML;
         $urlData = [
             "cms/storage-files/remote-upload",
         ];
-
-        if ($this->model)
-        {
-            $urlData = ArrayHelper::merge($urlData, [
-                "linked_to_model"   => $this->model->getRef()->getCode(),
-                "linked_to_value"   => $this->model->getRef()->getValue(),
-            ]);
-        }
-
 
         if ($this->fileGroup)
         {