+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 2.6.1.pre
 -----------------
  * Optimized widget content items
- * Added option to obtain all the descendants of the section element
+ * Added option to obtain all the descendants of the section element http://en.cms.skeeks.com/docs/sections-tree
  * Fixed critical bug with memory consumption when displaying files in the repository. There was at moments showing a large number of elements.
  * Revision of validation of additional properties + added examples: http://en.cms.skeeks.com/docs/additional-properties-models
  * Caching data tree to build the select element